src/processor/requirements.txt
azure-storage-queue
azure-storage-blob
azure-data-tables
pillow